games/requirements.txt
Django==5.2.5
djangorestframework==4.3.0
djangorestframework-simplejwt==6.4.0
